+++ src/ScottPlot5/ScottPlot5/Plot.cs
@@ -535,6 +535,26 @@ public class Plot : IDisposable
     }
 
     /// <summary>
+    /// Hide the default legend (inside the data area) and create a new legend panel 
+    /// placed on the edge of the figure outside the data area.
+    /// </summary>
+    /// <returns></returns>
+    public Panels.LegendPanel ShowLegend(Edge edge, Alignment alignment)
+    {
+        HideLegend();
+
+        Panels.LegendPanel legendPanel = new(Legend)
+        {
+            Edge = edge,
+            Alignment = alignment,
+        };
+
+        Axes.AddPanel(legendPanel);
+
+        return legendPanel;
+    }
+
+    /// <summary>
     /// Helper method for setting visibility of the <see cref="Legend"/>
     /// </summary>
     public Legend HideLegend()