+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 ## ScottPlot 5.0.35
-_Not yet on NuGet..._
+* Legend: Added `Plot.ShowLegend()` overload that accepts an `Edge` for quickly adding a legend outside the data area (#3672, #3635)
 
 ## ScottPlot 5.0.34
 _Published on [NuGet](https://www.nuget.org/profiles/ScottPlot) on 2024-05-05_